Add files via upload
The EFIS panel may not be fully initialized when rebooting the PC or after reconnecting it to an USB port. Everything seems to be fine but serial COMs are in a strange state.
As workaround you may have to soft reset the EFIS panel by pressing FPV and MTRS buttons at the same time. The backlight should blink 5 times.
You will have to perform this soft reset only once before starting SPAD, but every time the PC or the EFIS Panel are restarted.
Then when SPAD is started, and if the panel is fully ok, pressing both buttons will only turn ON and OFF the backlight without resetting it.